requirements.txt
python-telegram-bot==13.15
requests==2.31.0
python-dotenv==1.0.1
tzdata>=2023.3